Stocks/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
yfinance
pydantic
google-genai
